Refactor players page.
